+++ packages/client-web/src/store/index.ts
@@ -5,7 +5,7 @@ import { composeWithDevTools } from "redux-devtools-extension";
 import user from "./user";
 import date from "./date";
 import feeds from "./feeds";
-import timeline from "../modules/timeline/model";
+import timeline from "../modules/timeline/model/store";
 
 const reducer = combineReducers({
   user,
